requirements.txt
pyyaml>=6.0.1
openai>=1.14.0
matplotlib>=3.8.0
